\name{modtools-package}
\alias{modtools-package}
\alias{modtools}
\docType{package}
\title{
Additional tools for model diagnostic and selection
}
\description{
This package contains exhaustive regression functions and some tools
to evaluate the goodness of fit and the model quality.
}
\details{
\tabular{ll}{
Package: \tab modtools\cr
Type: \tab Package\cr
Version: \tab 1.0-7\cr
Date: \tab 2009-01-14\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.9.0),boot\cr
Suggests: \tab car\cr
}
# nothing for the moment
}
\author{
Pierre Bady <pierre.bady@free.fr>\cr
Maintainer: Pierre Bady <pierre.bady@free.fr>\cr
}
\references{
Saporta (2006)\cr
Efron et al \cr
Hinkley and Davidson (1998)\cr
etc\cr
to be completed\cr
\cr
}
\seealso{
\code{\link[car:car-package]{car}},
\code{\link[boot:boot-package]{boot}}
}
\examples{
# nothing for the moment
}
